// src/webparts/tiles/components/enums.ts
export enum ViewType {
  square = "square",
  circle = "circle",
}

export enum ColumnTypes {
  columns1 = "columns1",
  columns2 = "columns2",
  columns3 = "columns3",
  columns4 = "columns4",
}

export enum titlePosition {
  left = "left",
  center = "center",
  right = "right",
}

export enum textPosition {
  left = "left",
  bottom = "bottom",
  right = "right",
  overlay = "overlay",
  // overlayTransparent = "overlayTransparent",
}

export enum ActionType {
  create = "create",
  update = "update",
}